Update Binary Trees
I have calculated diameter of a tree in this code. 
I have made binary tree using structure taking each node as struct nod here, with its data and pointers to its left and right children. For calculating diameter I have used maxx as a parameter and initialized it to 0. At every node, I have calculated the height of its left and right subtrees and added them with 1(for root) to check if its greater than maxx, if it is then its value gets updated, else, maxx remains the same, and vice-versa for all nodes.